requirements.txt
numpy
opencv-python
torch
torchvision
Pillow
efficientnet_pytorch
scikit-learn
tqdm